refactor: status counter stage click
