+++ src/hooks/useStages.jsx
@@ -4,6 +4,7 @@ import { initialStages } from "../utils/StagesData";
 import { fetchCounts, filterCounter } from "../utils/statusCounterFunctions";
 import statusCounterStore from "../context/statusCounterStore";
 import { filterApplicants } from "../utils/applicantDataUtils";
+import moment from "moment";
 
 export const useStages = () => {
   //const [stages, setStages] = useState(initialStages);
@@ -61,3 +62,44 @@ export const useStages = () => {
 
   return { stages, setStages, toggleStage, toggleStatus };
 };
+
+export const handleStageClick = (stage, setSelectedStatuses, search, toggleStage, dateFilterType, dateFilter, positionFilter, setApplicantData, searchApplicant) => {
+    const stageStatuses = stage.statuses.map((status) => status.value);
+
+    setSelectedStatuses((prevStatuses) => {
+      // Check if all statuses in the stage are already selected
+      const allSelected = stageStatuses.every((status) =>
+        prevStatuses.includes(status),
+      );
+
+      let updatedStatuses;
+      if (allSelected) {
+        // If all statuses are selected, remove them
+        updatedStatuses = prevStatuses.filter(
+          (status) => !stageStatuses.includes(status),
+        );
+      } else {
+        // Otherwise, add the statuses that are not already selected
+        updatedStatuses = [
+          ...prevStatuses,
+          ...stageStatuses.filter((status) => !prevStatuses.includes(status)),
+        ];
+      }
+
+      // Call filterApplicants with the updated statuses
+      //filterApplicants(positionFilter, setApplicantData, updatedStatuses, moment(dateFilter).format("MMMM"), dateFilterType);
+      if (search === "") {
+        dateFilterType === 'month' ?
+        filterApplicants(positionFilter, setApplicantData, updatedStatuses, moment(dateFilter).format("MMMM"), dateFilterType) :
+        filterApplicants(positionFilter, setApplicantData, updatedStatuses, moment(dateFilter).format("YYYY"), dateFilterType)
+      }
+      else {
+        searchApplicant(search, setApplicantData, positionFilter, updatedStatuses, dateFilterType, dateFilter);
+      }
+
+      return updatedStatuses; // Return the updated state
+    });
+
+    // Toggle the stage's selected state
+    toggleStage(stage.name);
+  };
